@@ -4,12 +4,17 @@ import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.egov.common.data.query.builder.SelectQueryBuilder;
 import org.egov.common.data.query.exception.QueryBuilderException;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.models.core.SearchResponse;
 import org.egov.common.producer.Producer;
+import org.egov.common.utils.CommonUtils;
+import org.egov.common.utils.MultiStateInstanceUtil;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.util.ObjectUtils;
 import org.springframework.util.ReflectionUtils;
 
 import java.lang.reflect.Method;
@@ -28,6 +33,8 @@ import static org.egov.common.utils.CommonUtils.constructTotalCountCTEAndReturnR
 import static org.egov.common.utils.CommonUtils.getIdMethod;
 import static org.egov.common.utils.CommonUtils.getMethod;
 import static org.egov.common.utils.CommonUtils.getObjClass;
+import static org.egov.common.utils.CommonUtils.getSchemaName;
+import static org.egov.common.utils.MultiStateInstanceUtil.SCHEMA_REPLACE_STRING;
 
 /**
  * Generic Repository Class for common data operations.
@@ -47,6 +54,9 @@ public abstract class GenericRepository<T> {
 
     protected final RowMapper<T> rowMapper;
 
+    @Autowired
+    protected MultiStateInstanceUtil multiStateInstanceUtil;
+
     protected String tableName;
 
     @Value("${spring.cache.redis.time-to-live:60}")
@@ -65,29 +75,35 @@ public abstract class GenericRepository<T> {
     }
 
     /**
-     * Finds entities by their IDs.
+     * Finds entities by their IDs for given tenant
      *
+     * @param tenantId tenant id to find for.
      * @param ids The list of IDs to search for.
      * @return A list of entities found by the given IDs.
      */
-    public List<T> findById(List<String> ids) {
-        return findById(ids, false);
+    public List<T> findById(String tenantId, List<String> ids) throws InvalidTenantIdException {
+        return findById(tenantId, ids, false);
     }
 
     /**
      * Finds entities by their IDs with an option to include deleted entities.
      *
+     * @param tenantId       The tenant id to search for.
      * @param ids            The list of IDs to search for.
-     * @param includeDeleted Flag to include deleted entities in the search result.
      * @return A list of entities found by the given IDs.
      */
-    protected List<T> findInCache(List<String> ids) {
+    protected List<T> findInCache(String tenantId, List<String> ids) {
         ArrayList<T> objFound = new ArrayList<>();
         Collection<Object> collection = ids.stream().filter(Objects::nonNull)
                 .collect(Collectors.toList());
         log.info("Searching in cache");
+        String cacheTableName = tableName;
+        String schemaName = getSchemaName(tenantId, multiStateInstanceUtil);
+        if(!ObjectUtils.isEmpty(schemaName)) {
+            cacheTableName = schemaName.concat("-").concat(cacheTableName);
+        }
         List<Object> objFromCache = redisTemplate.opsForHash()
-                .multiGet(tableName, collection).stream().filter(Objects::nonNull).collect(Collectors.toList());
+                .multiGet(cacheTableName, collection).stream().filter(Objects::nonNull).collect(Collectors.toList());
         if (!objFromCache.isEmpty()) {
             if (objFromCache.size() == 1 && objFromCache.contains(null)) {
                 log.info("Cache miss");
@@ -106,25 +122,27 @@ public abstract class GenericRepository<T> {
      * Finds entities by their IDs with an option to include deleted entities,
      * using the default column name "id" for ID search.
      *
+     * @param tenantId       tenant id to search for.
      * @param ids            The list of IDs to search for.
      * @param includeDeleted Flag to include deleted entities in the search result.
      * @return A list of entities found by the given IDs.
      */
-    public List<T> findById(List<String> ids, Boolean includeDeleted) {
+    public List<T> findById(String tenantId, List<String> ids, Boolean includeDeleted) throws InvalidTenantIdException {
         // Delegates to the main findById method with the default column name "id"
-        return findById(ids, includeDeleted, "id");
+        return findById(tenantId, ids, includeDeleted, "id");
     }
 
     /**
      * Finds entities by their IDs with options to include deleted entities and specify a column name.
      *
+     * @param tenantId       tenant id to search for.
      * @param ids            The list of IDs to search for.
      * @param includeDeleted Flag to include deleted entities in the search result.
      * @param columnName     The name of the column to search IDs in.
      * @return A list of entities found by the given IDs.
      */
-    public List<T> findById(List<String> ids, Boolean includeDeleted, String columnName) {
-        List<T> objFound = findInCache(ids);
+    public List<T> findById(String tenantId, List<String> ids, Boolean includeDeleted, String columnName) throws InvalidTenantIdException {
+        List<T> objFound = findInCache(tenantId, ids);
 
         if (!objFound.isEmpty()) {
             Method idMethod = getIdMethod(objFound, columnName);
@@ -142,13 +160,13 @@ public abstract class GenericRepository<T> {
             }
         }
 
-        String query = String.format("SELECT * FROM %s WHERE %s IN (:ids) AND isDeleted = false", tableName, columnName);
+        String query = String.format("SELECT * FROM %s.%s WHERE %s IN (:ids) AND isDeleted = false", SCHEMA_REPLACE_STRING, tableName, columnName);
         if (null != includeDeleted && includeDeleted) {
-            query = String.format("SELECT * FROM %s WHERE %s IN (:ids)", tableName, columnName);
+            query = String.format("SELECT * FROM %s.%s WHERE %s IN (:ids)", SCHEMA_REPLACE_STRING, tableName, columnName);
         }
         Map<String, Object> paramMap = new HashMap<>();
         paramMap.put("ids", ids);
-
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
         objFound.addAll(namedParameterJdbcTemplate.query(query, paramMap, rowMapper));
         putInCache(objFound);
 
@@ -163,12 +181,15 @@ public abstract class GenericRepository<T> {
      * @return The list of saved entities.
      */
     public List<T> save(List<T> objects, String topic) {
-        producer.push(topic, objects);
+        String tenantId = CommonUtils.getTenantId(objects);
+        producer.push(tenantId, topic, objects);
         log.info("Pushed to kafka");
         putInCache(objects);
         log.info("Saved to cache");
         return objects;
     }
+
+
 
     /**
      * Saves entities to Kafka, caches them with specified cache key.
@@ -179,7 +200,8 @@ public abstract class GenericRepository<T> {
      * @return The list of saved entities.
      */
     public List<T> save(List<T> objects, String topic, String cacheKey) {
-        producer.push(topic, objects);
+        String tenantId = CommonUtils.getTenantId(objects);
+        producer.push(tenantId, topic, objects);
         log.info("Pushed to kafka");
         putInCache(objects, cacheKey);
         log.info("Saved to cache");
@@ -202,8 +224,14 @@ public abstract class GenericRepository<T> {
                                         obj -> obj,
                                         // in case of duplicates pick the latter
                                         (obj1, obj2) -> obj2));
-                redisTemplate.opsForHash().putAll(tableName, objMap);
-                redisTemplate.expire(tableName, Long.parseLong(timeToLive), TimeUnit.SECONDS);
+                String cacheTableName = tableName;
+                String tenantId = CommonUtils.getTenantId(objects);
+                String schemaName = getSchemaName(tenantId, multiStateInstanceUtil);
+                if(!ObjectUtils.isEmpty(schemaName)) {
+                    cacheTableName = schemaName.concat("-").concat(cacheTableName);
+                }
+                redisTemplate.opsForHash().putAll(cacheTableName, objMap);
+                redisTemplate.expire(cacheTableName, Long.parseLong(timeToLive), TimeUnit.SECONDS);
             }
         } catch (Exception exception) {
             log.warn("Error while saving to cache: {}", ExceptionUtils.getStackTrace(exception));
@@ -255,8 +283,8 @@ public abstract class GenericRepository<T> {
                                            Integer offset,
                                            String tenantId,
                                            Long lastChangedSince,
-                                           Boolean includeDeleted) throws QueryBuilderException {
-        String query = selectQueryBuilder.build(searchObject, tableName);
+                                           Boolean includeDeleted) throws QueryBuilderException, InvalidTenantIdException {
+        String query = selectQueryBuilder.build(searchObject, tableName, SCHEMA_REPLACE_STRING);
         query += " AND tenantId=:tenantId ";
         if (query.contains(tableName + " AND")) {
             query = query.replace(tableName + " AND", tableName + " WHERE");
@@ -272,7 +300,7 @@ public abstract class GenericRepository<T> {
         paramsMap.put("tenantId", tenantId);
         paramsMap.put("isDeleted", includeDeleted);
         paramsMap.put("lastModifiedTime", lastChangedSince);
-
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
         Long totalCount = constructTotalCountCTEAndReturnResult(query, paramsMap, namedParameterJdbcTemplate);
 
         query += " LIMIT :limit OFFSET :offset";
@@ -301,8 +329,8 @@ public abstract class GenericRepository<T> {
                         Integer offset,
                         String tenantId,
                         Long lastChangedSince,
-                        Boolean includeDeleted) throws QueryBuilderException {
-        String query = selectQueryBuilder.build(searchObject, tableName);
+                        Boolean includeDeleted) throws QueryBuilderException, InvalidTenantIdException {
+        String query = selectQueryBuilder.build(searchObject, tableName, SCHEMA_REPLACE_STRING);
         query += " AND tenantId=:tenantId ";
         if (query.contains(tableName + " AND")) {
             query = query.replace(tableName + " AND", tableName + " WHERE");
@@ -320,18 +348,20 @@ public abstract class GenericRepository<T> {
         paramsMap.put("lastModifiedTime", lastChangedSince);
         paramsMap.put("limit", limit);
         paramsMap.put("offset", offset);
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
         return namedParameterJdbcTemplate.query(query, paramsMap, rowMapper);
     }
 
     /**
      * Validates IDs against existing entities.
      *
+     * @param tenantId      tenant id to validate for.
      * @param idsToValidate The list of IDs to validate.
      * @param columnName    The name of the column containing IDs.
      * @return A list of valid IDs.
      */
-    public List<String> validateIds(List<String> idsToValidate, String columnName){
-        List<T> validIds = findById(idsToValidate, false, columnName);
+    public List<String> validateIds(String tenantId, List<String> idsToValidate, String columnName) throws InvalidTenantIdException {
+        List<T> validIds = findById(tenantId, idsToValidate, false, columnName);
         if (validIds.isEmpty()) {
             return Collections.emptyList();
         }
@@ -340,20 +370,20 @@ public abstract class GenericRepository<T> {
                 .collect(Collectors.toList());
     }
 
-    public List<String> validateClientReferenceIdsFromDB(List<String> clientReferenceIds, Boolean isDeletedKeyPresent) {
+    public List<String> validateClientReferenceIdsFromDB(String tenantId, List<String> clientReferenceIds, Boolean isDeletedKeyPresent) throws InvalidTenantIdException {
         List<String> objFound = new ArrayList<>();
 
         String query = null;
 
         if(isDeletedKeyPresent) {
-            query = String.format("SELECT clientReferenceId FROM %s WHERE clientReferenceId IN (:ids) AND isDeleted = false", tableName);
+            query = String.format("SELECT clientReferenceId FROM %s.%s WHERE clientReferenceId IN (:ids) AND isDeleted = false", SCHEMA_REPLACE_STRING, tableName);
         } else {
-            query = String.format("SELECT clientReferenceId FROM %s WHERE clientReferenceId IN (:ids) ", tableName);
+            query = String.format("SELECT clientReferenceId FROM %s.%s WHERE clientReferenceId IN (:ids) ", SCHEMA_REPLACE_STRING, tableName);
         }
 
         Map<String, Object> paramMap = new HashMap<>();
         paramMap.put("ids", clientReferenceIds);
-
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
         objFound.addAll(namedParameterJdbcTemplate.queryForList(query, paramMap, String.class));
 
         return objFound;
