@@ -1,7 +1,8 @@
 package org.egov.hrms.model;
 
 import lombok.*;
-import org.hibernate.validator.constraints.SafeHtml;
+import org.owasp.html.HtmlPolicyBuilder;
+import org.owasp.html.PolicyFactory;
 import org.springframework.validation.annotation.Validated;
 
 import javax.validation.constraints.NotNull;
@@ -16,30 +17,45 @@ import javax.validation.constraints.NotNull;
 @Builder
 public class ReactivationDetails {
 
-	@SafeHtml
+	private static final PolicyFactory POLICY = new HtmlPolicyBuilder().toFactory();
+
 	private String id;
 
-	@SafeHtml
 	@NotNull
 	private String reasonForReactivation;
 
-	@SafeHtml
 	private String orderNo;
 
-	@SafeHtml
 	private String remarks;
 
 	@NotNull
 	private Long effectiveFrom;
 
-	@SafeHtml
 	private String tenantId;
 
 	private AuditDetails auditDetails;
 
+	public void setId(String id) {
+		this.id = sanitize(id);
+	}
 
+	public void setReasonForReactivation(String reasonForReactivation) {
+		this.reasonForReactivation = sanitize(reasonForReactivation);
+	}
 
+	public void setOrderNo(String orderNo) {
+		this.orderNo = sanitize(orderNo);
+	}
 
+	public void setRemarks(String remarks) {
+		this.remarks = sanitize(remarks);
+	}
+
+	public void setTenantId(String tenantId) {
+		this.tenantId = sanitize(tenantId);
+	}
+
+	private String sanitize(String input) {
+		return input == null ? null : POLICY.sanitize(input);
+	}
 }
-
-
