@@ -1,51 +1,11 @@
-/*
- * eGov suite of products aim to improve the internal efficiency,transparency,
- * accountability and the service delivery of the government  organizations.
- *
- *  Copyright (C) 2016  eGovernments Foundation
- *
- *  The updated version of eGov suite of products as by eGovernments Foundation
- *  is available at http://www.egovernments.org
- *
- *  This program is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  any later version.
- *
- *  This program is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with this program. If not, see http://www.gnu.org/licenses/ or
- *  http://www.gnu.org/licenses/gpl.html .
- *
- *  In addition to the terms of the GPL license to be adhered to in using this
- *  program, the following additional terms are to be complied with:
- *
- *      1) All versions of this program, verbatim or modified must carry this
- *         Legal Notice.
- *
- *      2) Any misrepresentation of the origin of the material is prohibited. It
- *         is required that all modified versions of this material be marked in
- *         reasonable ways as different from the original version.
- *
- *      3) This license does not grant any rights to any user of the program
- *         with regards to rights under trademark law for use of the trade names
- *         or trademarks of eGovernments Foundation.
- *
- *  In case of any queries, you can reach eGovernments Foundation at contact@egovernments.org.
- */
-
 package org.egov.hrms.model;
 
 import lombok.*;
+import org.owasp.html.HtmlPolicyBuilder;
+import org.owasp.html.PolicyFactory;
+import org.springframework.validation.annotation.Validated;
 
 import javax.validation.constraints.NotNull;
-
-import org.hibernate.validator.constraints.SafeHtml;
-import org.springframework.validation.annotation.Validated;
 
 @Validated
 @EqualsAndHashCode(exclude = {"auditDetails"})
@@ -57,29 +17,47 @@ import org.springframework.validation.annotation.Validated;
 @ToString
 public class ServiceHistory {
 
-	@SafeHtml
+	private static final PolicyFactory POLICY = new HtmlPolicyBuilder().toFactory();
+
 	private String id;
 
-	@SafeHtml
 	private String serviceStatus;
 
 	private Long serviceFrom;
 
 	private Long serviceTo;
 
-	@SafeHtml
 	private String orderNo;
 
-	@SafeHtml
 	private String location;
 
-	@SafeHtml
-	private String tenantId;	
+	private String tenantId;
 
-	private  Boolean isCurrentPosition;
+	private Boolean isCurrentPosition;
 
 	private AuditDetails auditDetails;
 
+	public void setId(String id) {
+		this.id = sanitize(id);
+	}
 
+	public void setServiceStatus(String serviceStatus) {
+		this.serviceStatus = sanitize(serviceStatus);
+	}
 
+	public void setOrderNo(String orderNo) {
+		this.orderNo = sanitize(orderNo);
+	}
+
+	public void setLocation(String location) {
+		this.location = sanitize(location);
+	}
+
+	public void setTenantId(String tenantId) {
+		this.tenantId = sanitize(tenantId);
+	}
+
+	private String sanitize(String input) {
+		return input == null ? null : POLICY.sanitize(input);
+	}
 }
