@@ -13,6 +13,8 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+import static org.egov.common.utils.MultiStateInstanceUtil.SCHEMA_REPLACE_STRING;
+
 @Service
 public class EmployeeQueryBuilder {
 	
@@ -30,12 +32,15 @@ public class EmployeeQueryBuilder {
 	 */
 	public String getEmployeeSearchQuery(EmployeeSearchCriteria criteria,List <Object> preparedStmtList, Boolean addPagination ) {
 		StringBuilder builder = new StringBuilder(EmployeeQueries.HRMS_GET_EMPLOYEES);
+		builder = new StringBuilder(String.format(String.valueOf(builder), SCHEMA_REPLACE_STRING, SCHEMA_REPLACE_STRING ,SCHEMA_REPLACE_STRING, SCHEMA_REPLACE_STRING, SCHEMA_REPLACE_STRING,
+				SCHEMA_REPLACE_STRING, SCHEMA_REPLACE_STRING, SCHEMA_REPLACE_STRING, SCHEMA_REPLACE_STRING));
 		addWhereClause(criteria, builder, preparedStmtList);
 		return paginationClause(criteria, builder, addPagination);
 	}
 
 	public String getEmployeeCountQuery(String tenantId, List <Object> preparedStmtList ) {
 		StringBuilder builder = new StringBuilder(EmployeeQueries.HRMS_COUNT_EMP_QUERY);
+		builder = new StringBuilder(String.format(String.valueOf(builder), SCHEMA_REPLACE_STRING));
 		if(tenantId.equalsIgnoreCase(properties.stateLevelTenantId)){
 			builder.append("LIKE ? ");
 			preparedStmtList.add(tenantId+"%");
@@ -49,7 +54,7 @@ public class EmployeeQueryBuilder {
 	}
 	
 	public String getPositionSeqQuery() {
-		return EmployeeQueries.HRMS_POSITION_SEQ;
+		return  String.format(EmployeeQueries.HRMS_POSITION_SEQ, SCHEMA_REPLACE_STRING) ;
 	}
 	
 	/**
@@ -117,6 +122,7 @@ public class EmployeeQueryBuilder {
 
 	public String getAssignmentSearchQuery(EmployeeSearchCriteria criteria, List<Object> preparedStmtList) {
 		StringBuilder builder = new StringBuilder(EmployeeQueries.HRMS_GET_ASSIGNMENT);
+		builder = new StringBuilder(String.format(String.valueOf(builder), SCHEMA_REPLACE_STRING));
 		addWhereClauseAssignment(criteria, builder, preparedStmtList);
 		return builder.toString();
 	}
@@ -164,6 +170,7 @@ public class EmployeeQueryBuilder {
 
 	public String getUnassignedEmployeesSearchQuery(EmployeeSearchCriteria criteria, List<Object> preparedStmtList) {
 		StringBuilder builder = new StringBuilder(EmployeeQueries.HRMS_GET_UNASSIGNED_EMPLOYEES);
+		builder = new StringBuilder(String.format(String.valueOf(builder), SCHEMA_REPLACE_STRING, SCHEMA_REPLACE_STRING));
 		addWhereClauseAssignment(criteria, builder, preparedStmtList);
 		return builder.toString();
 	}
