@@ -7,8 +7,12 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import org.egov.common.contract.request.RequestInfo;
+import org.egov.common.exception.InvalidTenantIdException;
+import org.egov.common.utils.MultiStateInstanceUtil;
+import org.egov.hrms.utils.ErrorConstants;
 import org.egov.hrms.utils.HRMSUtils;
 import org.egov.hrms.web.contract.EmployeeSearchCriteria;
+import org.egov.tracer.model.CustomException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
@@ -21,7 +25,6 @@ import org.springframework.util.CollectionUtils;
 @Repository
 @Slf4j
 public class EmployeeRepository {
-
 	@Autowired
 	private EmployeeQueryBuilder queryBuilder;
 	
@@ -36,6 +39,9 @@ public class EmployeeRepository {
 
 	@Autowired
 	private HRMSUtils hrmsUtils;
+
+	@Autowired
+	private MultiStateInstanceUtil multiStateInstanceUtil;
 	
 	/**
 	 * DB Repository that makes jdbc calls to the db and fetches employees.
@@ -46,6 +52,7 @@ public class EmployeeRepository {
 	 */
 	public Map<String, Object> fetchEmployees(EmployeeSearchCriteria criteria, RequestInfo requestInfo){
 		Long totalCount = 0L;
+		String tenantId = criteria.getTenantId();
 		List<Employee> employees = new ArrayList<>();
 		List<Object> preparedStmtList = new ArrayList<>();
 		List<Object> preparedStmtListWithOutLimitAndOffset = new ArrayList<>();
@@ -69,7 +76,14 @@ public class EmployeeRepository {
 			criteria.setUuids(empUuids);
 		}
 		String query = queryBuilder.getEmployeeSearchQuery(criteria, preparedStmtList, true);
+
 		String queryWithOutLimitAndOffset = queryBuilder.getEmployeeSearchQuery(criteria, preparedStmtListWithOutLimitAndOffset , false);
+		try {
+			query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
+			queryWithOutLimitAndOffset = multiStateInstanceUtil.replaceSchemaPlaceholder(queryWithOutLimitAndOffset, tenantId);
+		} catch (InvalidTenantIdException e) {
+			throw new CustomException(ErrorConstants.TENANT_ID_EXCEPTION, e.getMessage());
+		}
 		try {
 			employees = jdbcTemplate.query(query, preparedStmtList.toArray(),rowMapper);
 			totalCount = jdbcTemplate.query(queryWithOutLimitAndOffset, preparedStmtList.toArray(),rowMapper).spliterator().getExactSizeIfKnown();
@@ -85,7 +99,13 @@ public class EmployeeRepository {
 	private List<String> fetchUnassignedEmployees(EmployeeSearchCriteria criteria, RequestInfo requestInfo) {
 		List<String> employeesIds = new ArrayList<>();
 		List <Object> preparedStmtList = new ArrayList<>();
+		String tenantId = criteria.getTenantId();
 		String query = queryBuilder.getUnassignedEmployeesSearchQuery(criteria, preparedStmtList);
+		try {
+			query = multiStateInstanceUtil.replaceSchemaPlaceholder(query,tenantId);
+		} catch (InvalidTenantIdException e) {
+			throw new CustomException(ErrorConstants.TENANT_ID_EXCEPTION, e.getMessage());
+		}
 		try {
 			employeesIds = jdbcTemplate.queryForList(query, preparedStmtList.toArray(),String.class);
 		}catch(Exception e) {
@@ -98,7 +118,14 @@ public class EmployeeRepository {
 	private List<String> fetchEmployeesforAssignment(EmployeeSearchCriteria criteria, RequestInfo requestInfo) {
 		List<String> employeesIds = new ArrayList<>();
 		List <Object> preparedStmtList = new ArrayList<>();
+		String tenantId = criteria.getTenantId();
 		String query = queryBuilder.getAssignmentSearchQuery(criteria, preparedStmtList);
+		try {
+			query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
+		} catch (InvalidTenantIdException e) {
+			throw new CustomException(ErrorConstants.TENANT_ID_EXCEPTION, e.getMessage());
+		}
+
 		try {
 
 			employeesIds = jdbcTemplate.queryForList(query, preparedStmtList.toArray(),String.class);
@@ -115,9 +142,14 @@ public class EmployeeRepository {
 	 * 
 	 * @return
 	 */
-	public Long fetchPosition(){
+	public Long fetchPosition(String tenantId){
 		String query = queryBuilder.getPositionSeqQuery();
 		Long id = null;
+		try {
+			query = multiStateInstanceUtil.replaceSchemaPlaceholder( query, tenantId);
+		} catch (InvalidTenantIdException e) {
+			throw new CustomException(ErrorConstants.TENANT_ID_EXCEPTION, e.getMessage());
+		}
 		try {
 			id = jdbcTemplate.queryForObject(query, Long.class);
 		}catch(Exception e) {
@@ -139,7 +171,12 @@ public class EmployeeRepository {
 
 		String query = queryBuilder.getEmployeeCountQuery(tenantId, preparedStmtList);
 		log.info("query; "+query);
-		try {
+        try {
+			query = multiStateInstanceUtil.replaceSchemaPlaceholder( query, tenantId);
+        } catch (InvalidTenantIdException e) {
+            throw new CustomException(ErrorConstants.TENANT_ID_EXCEPTION, e.getMessage());
+        }
+        try {
 			response=jdbcTemplate.query(query, preparedStmtList.toArray(),countRowMapper);
 		}catch(Exception e) {
 			log.error("Exception while making the db call: ",e);
