@@ -118,10 +118,11 @@ public class EmployeeService {
 	 */
 	public EmployeeResponse create(EmployeeRequest employeeRequest) {
 		RequestInfo requestInfo = employeeRequest.getRequestInfo();
+		String tenantId = employeeRequest.getEmployees().stream().findAny().get().getTenantId();
 		Map<String, String> pwdMap = new HashMap<>();
 		idGenService.setIds(employeeRequest);
 		employeeRequest.getEmployees().stream().forEach(employee -> {
-			enrichCreateRequest(employee, requestInfo);
+			enrichCreateRequest(tenantId, employee, requestInfo);
 			createUser(employee, requestInfo);
 			pwdMap.put(employee.getUuid(), employee.getUser().getPassword());
 		});
@@ -130,7 +131,7 @@ public class EmployeeService {
 		// Setting password as null after sending employeeRequest to email notification topic to send email.
 		employeeRequest.getEmployees().forEach(employee -> employee.getUser().setPassword(null));
 
-		hrmsProducer.push(propertiesManager.getSaveEmployeeTopic(), employeeRequest);
+		hrmsProducer.push(tenantId, propertiesManager.getSaveEmployeeTopic(), employeeRequest);
 		notificationService.sendNotification(employeeRequest, pwdMap);
 		return generateResponse(employeeRequest);
 	}
@@ -228,8 +229,6 @@ public class EmployeeService {
 					criteria.setUuids(userUUIDs);
 			}
 		}
-		if(userChecked)
-			criteria.setTenantId(null);
         List <Employee> employees = new ArrayList<>();
         if(!((!CollectionUtils.isEmpty(criteria.getRoles()) || !CollectionUtils.isEmpty(criteria.getNames()) || !StringUtils.isEmpty(criteria.getPhone())) && CollectionUtils.isEmpty(criteria.getUuids()))) {
 			Map<String, Object> response = repository.fetchEmployees(criteria, requestInfo);
@@ -238,6 +237,8 @@ public class EmployeeService {
 			totalCount = (Long) response.get("totalCount");
 		}
         List<String> uuids = employees.stream().map(Employee :: getUuid).collect(Collectors.toList());
+		if(userChecked)
+			criteria.setTenantId(null);
 		if(!CollectionUtils.isEmpty(uuids)){
             Map<String, Object> userSearchCriteria = new HashMap<>();
 			userSearchCriteria.put(HRMSConstants.HRMS_USER_SERACH_CRITERIA_USERTYPE_CODE, HRMSConstants.HRMS_USER_SERACH_CRITERIA_USERTYPE);
@@ -320,8 +321,7 @@ public class EmployeeService {
 	 * @param employee
 	 * @param requestInfo
 	 */
-	private void enrichCreateRequest(Employee employee, RequestInfo requestInfo) {
-
+	private void enrichCreateRequest(String tenantId, Employee employee, RequestInfo requestInfo) {
 		AuditDetails auditDetails = AuditDetails.builder()
 				.createdBy(requestInfo.getUserInfo().getUuid())
 				.createdDate(new Date().getTime())
@@ -335,9 +335,10 @@ public class EmployeeService {
 		});
 		if (employee.getAssignments() != null && !employee.getAssignments().isEmpty()) {
 			employee.getAssignments().stream().forEach(assignment -> {
+//				String tenantId = assignment.getTenantid();
 				assignment.setId(UUID.randomUUID().toString());
 				assignment.setAuditDetails(auditDetails);
-				assignment.setPosition(getPosition());
+				assignment.setPosition(getPosition(tenantId));
 			});
 		}
 		if(!CollectionUtils.isEmpty(employee.getServiceHistory())) {
@@ -378,8 +379,8 @@ public class EmployeeService {
 	 * Fetches next value from the position sequence table
 	 * @return
 	 */
-	public Long getPosition() {
-		return repository.fetchPosition();
+	public Long getPosition(String tenantId) {
+		return repository.fetchPosition(tenantId);
 	}
 
 	/**
@@ -429,7 +430,7 @@ public class EmployeeService {
 		});
 
 		// Push the updated employee request to the HRMS topic for further processing
-		hrmsProducer.push(propertiesManager.getUpdateTopic(), employeeRequest);
+		hrmsProducer.push(tenantId, propertiesManager.getUpdateTopic(), employeeRequest);
 
 		// (Optional) Send reactivation notifications if needed
 		// notificationService.sendReactivationNotification(employeeRequest);
