@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.*;
 
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import java.util.Set;
@@ -20,10 +21,10 @@ public class Email {
     @Size(min = 1, message = "At least one recipient is required")
     private Set<String> emailTo;
 
-    @NotBlank(message = "Subject is required")
+    @NotNull(message = "Subject is required")
     private String subject;
 
-    @NotBlank(message = "Body is required")
+    @NotNull(message = "Body is required")
     private String body;
     @JsonProperty("isHTML")
     private boolean isHTML;
