@@ -1,6 +1,7 @@
 package org.egov.household.household.member.validators;
 
 import lombok.extern.slf4j.Slf4j;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.models.Error;
 import org.egov.common.models.household.HouseholdMember;
 import org.egov.common.models.household.HouseholdMemberBulkRequest;
@@ -22,8 +23,8 @@ import java.util.stream.Collectors;
 import static org.egov.common.utils.CommonUtils.getIdMethod;
 import static org.egov.common.utils.CommonUtils.notHavingErrors;
 import static org.egov.common.utils.CommonUtils.populateErrorDetails;
-import static org.egov.household.Constants.HOUSEHOLD_ALREADY_HAS_HEAD;
-import static org.egov.household.Constants.HOUSEHOLD_ALREADY_HAS_HEAD_MESSAGE;
+import static org.egov.common.utils.ValidatorUtils.getErrorForInvalidTenantId;
+import static org.egov.household.Constants.*;
 import static org.egov.household.utils.CommonUtils.getHouseholdColumnName;
 
 @Component
@@ -67,22 +68,31 @@ public class HmHouseholdHeadValidator implements Validator<HouseholdMemberBulkRe
                                          HashMap<HouseholdMember, List<Error>> errorDetailsMap) {
 
         if(householdMember.getIsHeadOfHousehold()){
-            log.info("validating if household already has a head");
-            List<HouseholdMember> householdMembersHeadCheck = householdMemberRepository
-                    .findIndividualByHousehold((String) ReflectionUtils.invokeMethod(idMethod, householdMember),
-                            columnName).getResponse().stream().filter(HouseholdMember::getIsHeadOfHousehold)
-                    .collect(Collectors.toList());
 
-            if(!householdMembersHeadCheck.isEmpty()){
-                Error error = Error.builder().errorMessage(HOUSEHOLD_ALREADY_HAS_HEAD_MESSAGE)
-                        .errorCode(HOUSEHOLD_ALREADY_HAS_HEAD)
-                        .type(Error.ErrorType.NON_RECOVERABLE)
-                        .exception(new CustomException(HOUSEHOLD_ALREADY_HAS_HEAD,
-                                HOUSEHOLD_ALREADY_HAS_HEAD_MESSAGE))
-                        .build();
-                log.info("household already has a head, error: {}", error);
+            String tenantId = householdMember.getTenantId();
+            log.info("validating if household already has a head");
+
+            try {
+                List<HouseholdMember> householdMembersHeadCheck = householdMembersHeadCheck = householdMemberRepository.findIndividualByHousehold( tenantId, (String) ReflectionUtils.invokeMethod(idMethod, householdMember),
+                                columnName).getResponse().stream().filter(HouseholdMember::getIsHeadOfHousehold)
+                        .collect(Collectors.toList());
+
+
+                if(!householdMembersHeadCheck.isEmpty()){
+                    Error error = Error.builder().errorMessage(HOUSEHOLD_ALREADY_HAS_HEAD_MESSAGE)
+                            .errorCode(HOUSEHOLD_ALREADY_HAS_HEAD)
+                            .type(Error.ErrorType.NON_RECOVERABLE)
+                            .exception(new CustomException(HOUSEHOLD_ALREADY_HAS_HEAD,
+                                    HOUSEHOLD_ALREADY_HAS_HEAD_MESSAGE))
+                            .build();
+                    log.info("household already has a head, error: {}", error);
+                    populateErrorDetails(householdMember, error, errorDetailsMap);
+                }
+            } catch (InvalidTenantIdException exception) {
+                Error error = getErrorForInvalidTenantId(tenantId, exception);
                 populateErrorDetails(householdMember, error, errorDetailsMap);
             }
+
         }
     }
 }
