@@ -1,7 +1,9 @@
 package org.egov.household.household.member.validators;
 
+import io.swagger.models.auth.In;
 import lombok.extern.slf4j.Slf4j;
 import org.egov.common.contract.request.RequestInfo;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.models.Error;
 import org.egov.common.models.household.HouseholdMember;
 import org.egov.common.models.household.HouseholdMemberBulkRequest;
@@ -22,6 +24,7 @@ import java.util.stream.Collectors;
 import static org.egov.common.utils.CommonUtils.getTenantId;
 import static org.egov.common.utils.CommonUtils.notHavingErrors;
 import static org.egov.common.utils.CommonUtils.populateErrorDetails;
+import static org.egov.common.utils.ValidatorUtils.getErrorForInvalidTenantId;
 import static org.egov.household.Constants.INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD;
 import static org.egov.household.Constants.INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD_MESSAGE;
 import static org.egov.household.utils.ValidatorUtil.getHouseholdMembersWithNonNullIndividuals;
@@ -68,16 +71,21 @@ public class HmUniqueIndividualValidator implements Validator<HouseholdMemberBul
                     householdMember.setIndividualClientReferenceId(individual.getClientReferenceId());
 
                     log.info("finding individuals mappings in household member");
-                    List<HouseholdMember> individualSearchResult = householdMemberRepository
-                            .findIndividual(individual.getId()).getResponse();
-                    if(!individualSearchResult.isEmpty()) {
-                        Error error = Error.builder().errorMessage(INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD_MESSAGE)
-                                .errorCode(INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD)
-                                .type(Error.ErrorType.NON_RECOVERABLE)
-                                .exception(new CustomException(INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD,
-                                        INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD_MESSAGE))
-                                .build();
-                        log.info("found error in individual mapping {}", error);
+                    try {
+                        List<HouseholdMember> individualSearchResult = householdMemberRepository
+                                .findIndividual(tenantId, individual.getId()).getResponse();
+                        if(!individualSearchResult.isEmpty()) {
+                            Error error = Error.builder().errorMessage(INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD_MESSAGE)
+                                    .errorCode(INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD)
+                                    .type(Error.ErrorType.NON_RECOVERABLE)
+                                    .exception(new CustomException(INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD,
+                                            INDIVIDUAL_ALREADY_MEMBER_OF_HOUSEHOLD_MESSAGE))
+                                    .build();
+                            log.info("found error in individual mapping {}", error);
+                            populateErrorDetails(householdMember, error, errorDetailsMap);
+                        }
+                    } catch (InvalidTenantIdException exception) {
+                        Error error = getErrorForInvalidTenantId(tenantId, exception);
                         populateErrorDetails(householdMember, error, errorDetailsMap);
                     }
                 }
