@@ -6,6 +6,7 @@ import org.egov.common.data.query.builder.QueryFieldChecker;
 import org.egov.common.data.query.builder.SelectQueryBuilder;
 import org.egov.common.data.query.exception.QueryBuilderException;
 import org.egov.common.data.repository.GenericRepository;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.models.core.SearchResponse;
 import org.egov.common.models.household.Household;
 import org.egov.common.producer.Producer;
@@ -13,6 +14,7 @@ import org.egov.household.repository.rowmapper.HouseholdRowMapper;
 import org.egov.common.models.household.HouseholdSearch;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ScanCursor;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.CollectionUtils;
@@ -27,6 +29,7 @@ import java.util.stream.Collectors;
 
 import static org.egov.common.utils.CommonUtils.constructTotalCountCTEAndReturnResult;
 import static org.egov.common.utils.CommonUtils.getIdMethod;
+import static org.egov.common.utils.MultiStateInstanceUtil.SCHEMA_REPLACE_STRING;
 
 @Repository
 @Slf4j
@@ -43,8 +46,8 @@ public class HouseholdRepository extends GenericRepository<Household> {
         super(producer, namedParameterJdbcTemplate, redisTemplate, selectQueryBuilder, householdRowMapper, Optional.of("household h"));
     }
 
-    public SearchResponse<Household> findById(List<String> ids, String columnName, Boolean includeDeleted) {
-        List<Household> objFound = findInCache(ids);
+    public SearchResponse<Household> findById(String tenantId, List<String> ids, String columnName, Boolean includeDeleted) throws InvalidTenantIdException {
+        List<Household> objFound = findInCache( tenantId,ids);
         if (!includeDeleted) {
             objFound = objFound.stream()
                     .filter(entity -> entity.getIsDeleted().equals(false))
@@ -60,23 +63,23 @@ public class HouseholdRepository extends GenericRepository<Household> {
             }
         }
 
-        String query = String.format("SELECT *, a.id as aid,a.tenantid as atenantid, a.clientreferenceid as aclientreferenceid FROM household h LEFT JOIN address a ON h.addressid = a.id WHERE h.%s IN (:ids) AND isDeleted = false", columnName);
+        String query = String.format("SELECT *, a.id as aid,a.tenantid as atenantid, a.clientreferenceid as aclientreferenceid FROM %s.household h LEFT JOIN %s.address a ON h.addressid = a.id WHERE h.%s IN (:ids) AND isDeleted = false",SCHEMA_REPLACE_STRING , SCHEMA_REPLACE_STRING, columnName);
         if (null != includeDeleted && includeDeleted) {
-            query = String.format("SELECT *, a.id as aid,a.tenantid as atenantid, a.clientreferenceid as aclientreferenceid FROM household h LEFT JOIN address a ON h.addressid = a.id  WHERE h.%s IN (:ids)", columnName);
+            query = String.format("SELECT *, a.id as aid,a.tenantid as atenantid, a.clientreferenceid as aclientreferenceid FROM %s.household h LEFT JOIN %s.address a ON h.addressid = a.id  WHERE h.%s IN (:ids)", SCHEMA_REPLACE_STRING , SCHEMA_REPLACE_STRING, columnName);
         }
         Map<String, Object> paramMap = new HashMap();
         paramMap.put("ids", ids);
 
+        query=  multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
         Long totalCount = constructTotalCountCTEAndReturnResult(query, paramMap, this.namedParameterJdbcTemplate);
 
         objFound.addAll(this.namedParameterJdbcTemplate.query(query, paramMap, this.rowMapper));
-        putInCache(objFound);
+//        putInCache(objFound);
         return SearchResponse.<Household>builder().totalCount(totalCount).response(objFound).build();
     }
 
-    public SearchResponse<Household> find(HouseholdSearch searchObject, Integer limit, Integer offset, String tenantId, Long lastChangedSince, Boolean includeDeleted) {
-        String query = "SELECT *, a.id as aid,a.tenantid as atenantid, a.clientreferenceid as aclientreferenceid";
-        query += " FROM household h LEFT JOIN address a ON h.addressid = a.id";
+    public SearchResponse<Household> find(HouseholdSearch searchObject, Integer limit, Integer offset, String tenantId, Long lastChangedSince, Boolean includeDeleted) throws InvalidTenantIdException {
+        String query = String.format("SELECT *, a.id as aid,a.tenantid as atenantid, a.clientreferenceid as aclientreferenceid FROM %s.household h LEFT JOIN %s.address a ON h.addressid = a.id", SCHEMA_REPLACE_STRING, SCHEMA_REPLACE_STRING) ;
         Map<String, Object> paramsMap = new HashMap<>();
         List<String> whereFields = GenericQueryBuilder.getFieldsWithCondition(searchObject, QueryFieldChecker.isNotNull, paramsMap);
         query = GenericQueryBuilder.generateQuery(query, whereFields).toString();
@@ -103,7 +106,7 @@ public class HouseholdRepository extends GenericRepository<Household> {
         paramsMap.put("tenantId", tenantId);
         paramsMap.put("isDeleted", includeDeleted);
         paramsMap.put("lastModifiedTime", lastChangedSince);
-
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
         Long totalCount = constructTotalCountCTEAndReturnResult(query, paramsMap, this.namedParameterJdbcTemplate);
 
         query = query + "ORDER BY h.id ASC LIMIT :limit OFFSET :offset";
@@ -124,10 +127,10 @@ public class HouseholdRepository extends GenericRepository<Household> {
      *
      * Fetch all the household which falls under the radius provided using longitude and latitude provided.
      */
-    public SearchResponse<Household> findByRadius(HouseholdSearch searchObject, Integer limit, Integer offset, String tenantId, Boolean includeDeleted) throws QueryBuilderException {
-        String query = searchCriteriaWaypointQuery +
+    public SearchResponse<Household> findByRadius(HouseholdSearch searchObject, Integer limit, Integer offset, String tenantId, Boolean includeDeleted) throws QueryBuilderException, InvalidTenantIdException {
+        String query = String.format(searchCriteriaWaypointQuery +
                 "SELECT * FROM (SELECT h.*, a.*, a.id as aid,a.tenantid as atenantid, a.clientreferenceid as aclientreferenceid, " + calculateDistanceFromTwoWaypointsFormulaQuery + " \n" +
-                "FROM public.household h LEFT JOIN public.address a ON h.addressid = a.id AND h.tenantid = a.tenantid, cte_search_criteria_waypoint cte_scw ";
+                "FROM %s.household h LEFT JOIN %s.address a ON h.addressid = a.id AND h.tenantid = a.tenantid, cte_search_criteria_waypoint cte_scw ", SCHEMA_REPLACE_STRING, SCHEMA_REPLACE_STRING);
         Map<String, Object> paramsMap = new HashMap<>();
         List<String> whereFields = GenericQueryBuilder.getFieldsWithCondition(searchObject, QueryFieldChecker.isNotNull, paramsMap);
         query = GenericQueryBuilder.generateQuery(query, whereFields).toString();
@@ -155,6 +158,7 @@ public class HouseholdRepository extends GenericRepository<Household> {
         paramsMap.put("isDeleted", includeDeleted);
         paramsMap.put("distance", searchObject.getSearchRadius());
         query = query + " ORDER BY distance ASC";
+        query = multiStateInstanceUtil.replaceSchemaPlaceholder(query, tenantId);
         Long totalCount = constructTotalCountCTEAndReturnResult(query, paramsMap, this.namedParameterJdbcTemplate);
         query = query + " LIMIT :limit OFFSET :offset ";
         paramsMap.put("limit", limit);
