@@ -6,10 +6,12 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import lombok.extern.slf4j.Slf4j;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.models.Error;
 import org.egov.common.models.household.Household;
 import org.egov.common.models.household.HouseholdBulkRequest;
 import org.egov.common.models.household.HouseholdSearch;
+import org.egov.common.utils.CommonUtils;
 import org.egov.common.validator.Validator;
 import org.egov.household.repository.HouseholdRepository;
 import org.springframework.core.annotation.Order;
@@ -19,7 +21,7 @@ import org.springframework.util.StringUtils;
 
 import static org.egov.common.utils.CommonUtils.notHavingErrors;
 import static org.egov.common.utils.CommonUtils.populateErrorDetails;
-import static org.egov.common.utils.ValidatorUtils.getErrorForUniqueEntity;
+import static org.egov.common.utils.ValidatorUtils.*;
 
 /**
  * Validator class for checking the existence of entities with the given client reference IDs.
@@ -55,7 +57,7 @@ public class HExistentEntityValidator implements Validator<HouseholdBulkRequest,
     public Map<Household, List<Error>> validate(HouseholdBulkRequest request) {
         // Map to hold household entities and their associated error details
         Map<Household, List<Error>> errorDetailsMap = new HashMap<>();
-
+        String tenantId = CommonUtils.getTenantId(request.getHouseholds()); // Extract tenant ID from the request
         // Get the list of household entities from the request
         List<Household> entities = request.getHouseholds();
 
@@ -78,16 +80,26 @@ public class HExistentEntityValidator implements Validator<HouseholdBulkRequest,
         // Check if the client reference ID list is not empty before querying the database
         if (!CollectionUtils.isEmpty(clientReferenceIdList)) {
             // Query the repository to find existing entities by client reference IDs
-            List<String> existingClientReferenceIds =
-                    householdRepository.validateClientReferenceIdsFromDB(clientReferenceIdList, Boolean.TRUE);
 
-            // For each existing client reference ID, populate error details for uniqueness
-            existingClientReferenceIds.forEach(clientReferenceId -> {
-                // Get a predefined error object for unique entity validation
-                Error error = getErrorForUniqueEntity();
-                // Populate error details for the household entity associated with the client reference ID
-                populateErrorDetails(map.get(clientReferenceId), error, errorDetailsMap);
-            });
+            try {
+                List<String> existingClientReferenceIds =
+                        householdRepository.validateClientReferenceIdsFromDB(tenantId, clientReferenceIdList, Boolean.TRUE);
+
+                // For each existing client reference ID, populate error details for uniqueness
+                existingClientReferenceIds.forEach(clientReferenceId -> {
+                    // Get a predefined error object for unique entity validation
+                    Error error = getErrorForUniqueEntity();
+                    // Populate error details for the household entity associated with the client reference ID
+                    populateErrorDetails(map.get(clientReferenceId), error, errorDetailsMap);
+                });
+            } catch (InvalidTenantIdException exception) {
+                entities.forEach(household -> {
+                    // If an exception occurs, populate error details for the household entity
+                    Error error = getErrorForInvalidTenantId(tenantId, exception);
+                    populateErrorDetails(household, error, errorDetailsMap);
+                });
+            }
+
         }
 
         // Return the map containing household entities and their associated error details
