@@ -2,6 +2,7 @@ package org.egov.household.service;
 
 import org.egov.common.data.query.exception.QueryBuilderException;
 import org.egov.common.ds.Tuple;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.helper.RequestInfoTestBuilder;
 import org.egov.common.models.core.SearchResponse;
 import org.egov.common.models.household.Household;
@@ -38,39 +39,39 @@ class HouseholdFindTest {
 
     @Test
     @DisplayName("should search only by id if only id is present")
-    void shouldOnlySearchByIdIfOnlyIdIsPresent() throws QueryBuilderException {
+    void shouldOnlySearchByIdIfOnlyIdIsPresent() throws QueryBuilderException, InvalidTenantIdException {
         HouseholdSearchRequest householdSearchRequest = HouseholdSearchRequest.builder()
                 .requestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
                 .household(HouseholdSearch.builder().id(Collections.singletonList("some-id")).build()).build();
-        when(householdRepository.findById(anyList(), eq("id"), anyBoolean()))
+        when(householdRepository.findById( anyString(),  anyList(), eq("id"), anyBoolean()))
                 .thenReturn(SearchResponse.<Household>builder().build());
 
         householdService.search(householdSearchRequest.getHousehold(), 10, 0, "default",
                 null, false);
 
         verify(householdRepository, times(1))
-                .findById(anyList(), eq("id"), anyBoolean());
+                .findById( anyString(),anyList(), eq("id"), anyBoolean());
     }
 
     @Test
     @DisplayName("should search only by clientReferenceId if only clientReferenceId is present")
-    void shouldOnlySearchByClientReferenceIdIfOnlyClientReferenceIdIsPresent() throws QueryBuilderException {
+    void shouldOnlySearchByClientReferenceIdIfOnlyClientReferenceIdIsPresent() throws QueryBuilderException, InvalidTenantIdException {
         HouseholdSearchRequest householdSearchRequest = HouseholdSearchRequest.builder()
                 .requestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
                 .household(HouseholdSearch.builder().clientReferenceId(Collections.singletonList("some-id")).build()).build();
-        when(householdRepository.findById(anyList(), eq("clientReferenceId"), anyBoolean()))
+        when(householdRepository.findById(anyString(), anyList(), eq("clientReferenceId"), anyBoolean()))
                 .thenReturn(SearchResponse.<Household>builder().build());
 
         householdService.search(householdSearchRequest.getHousehold(), 10, 0, "default",
                 null, false);
 
-        verify(householdRepository, times(1)).findById(anyList(),
+        verify(householdRepository, times(1)).findById(anyString(), anyList(),
                 eq("clientReferenceId"), anyBoolean());
     }
 
     @Test
     @DisplayName("should not call findById if more search parameters are available")
-    void shouldNotCallFindByIdIfMoreParametersAreAvailable() throws QueryBuilderException {
+    void shouldNotCallFindByIdIfMoreParametersAreAvailable() throws QueryBuilderException, InvalidTenantIdException {
         HouseholdSearchRequest householdSearchRequest = HouseholdSearchRequest.builder()
                 .requestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
                 .household(HouseholdSearch.builder().id(Collections.singletonList("someid"))
@@ -82,12 +83,12 @@ class HouseholdFindTest {
                 "default", 0L, false);
 
         verify(householdRepository, times(0))
-                .findById(anyList(), anyString(), anyBoolean());
+                .findById( anyString(),anyList(), anyString(), anyBoolean());
     }
 
     @Test
     @DisplayName("should call find if more parameters are available")
-    void shouldCallFindIfMoreParametersAreAvailable() throws QueryBuilderException {
+    void shouldCallFindIfMoreParametersAreAvailable() throws QueryBuilderException, InvalidTenantIdException {
         HouseholdSearchRequest householdSearchRequest = HouseholdSearchRequest.builder()
                 .requestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
                 .household(HouseholdSearch.builder().id(Collections.singletonList("someid"))
