@@ -1,6 +1,7 @@
 package org.egov.household.service;
 
 import org.egov.common.ds.Tuple;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.models.core.SearchResponse;
 import org.egov.common.models.household.Household;
 import org.egov.common.models.household.HouseholdMemberBulkRequest;
@@ -42,14 +43,19 @@ class HouseholdMemberCreateEnrichmentTest {
     }
 
     private void mockHouseholdFindIds() {
-        when(householdService.findById(
-                any(List.class),
-                any(String.class),
-                any(Boolean.class)
-        )).thenReturn(SearchResponse.<Household>builder()
-                .response(Collections.singletonList(
-                Household.builder().id("some-household-id").clientReferenceId("some-client-ref-id").build()))
-                .build());
+        try {
+            when(householdService.findById(
+                    any(String.class),
+                    any(List.class),
+                    any(String.class),
+                    any(Boolean.class)
+            )).thenReturn(SearchResponse.<Household>builder()
+                    .response(Collections.singletonList(
+                    Household.builder().id("some-household-id").clientReferenceId("some-client-ref-id").build()))
+                    .build());
+        } catch (InvalidTenantIdException e) {
+            throw new RuntimeException(e);
+        }
     }
 
     @Test
