@@ -1,6 +1,7 @@
 package org.egov.household.service;
 
 import org.egov.common.data.query.exception.QueryBuilderException;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.helper.RequestInfoTestBuilder;
 import org.egov.common.models.core.SearchResponse;
 import org.egov.common.models.household.HouseholdMember;
@@ -37,25 +38,25 @@ class HouseholdMemberFindTest {
 
     @Test
     @DisplayName("should search only by id if only id is present")
-    void shouldOnlySearchByIdIfOnlyIdIsPresent() {
+    void shouldOnlySearchByIdIfOnlyIdIsPresent() throws InvalidTenantIdException {
         HouseholdMemberSearchRequest householdSearchRequest = HouseholdMemberSearchRequest.builder()
                 .requestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
                 .householdMemberSearch(HouseholdMemberSearch.builder()
                         .id(Collections.singletonList("some-id")).build()).build();
-        when(householdMemberRepository.findById(anyList(), eq("id"), anyBoolean()))
+        when(householdMemberRepository.findById( anyString(), anyList(), eq("id"), anyBoolean()))
                 .thenReturn(SearchResponse.<HouseholdMember>builder().build());
 
         householdMemberService.search(householdSearchRequest.getHouseholdMemberSearch(), 10, 0, "default",
                 null, false);
 
         verify(householdMemberRepository, times(1))
-                .findById(anyList(), eq("id"), anyBoolean());
+                .findById( anyString(), anyList(), eq("id"), anyBoolean());
     }
 
 
     @Test
     @DisplayName("should not call findById if more search parameters are available")
-    void shouldNotCallFindByIfIfMoreParametersAreAvailable() throws QueryBuilderException {
+    void shouldNotCallFindByIfIfMoreParametersAreAvailable() throws QueryBuilderException, InvalidTenantIdException {
         HouseholdMemberSearchRequest householdMemberSearchRequest = HouseholdMemberSearchRequest.builder()
                 .requestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
                 .householdMemberSearch(HouseholdMemberSearch.builder()
@@ -67,13 +68,13 @@ class HouseholdMemberFindTest {
                 "", 0L, false);
 
         verify(householdMemberRepository, times(0))
-                .findById(anyList(), anyString(), anyBoolean());
+                .findById(anyString() ,anyList(), anyString(), anyBoolean());
     }
 
 
     @Test
     @DisplayName("should call find if more parameters are available")
-    void shouldCallFindIfMoreParametersAreAvailable() throws QueryBuilderException {
+    void shouldCallFindIfMoreParametersAreAvailable() throws QueryBuilderException, InvalidTenantIdException {
         HouseholdMemberSearchRequest householdMemberSearchRequest = HouseholdMemberSearchRequest.builder()
                 .requestInfo(RequestInfoTestBuilder.builder().withCompleteRequestInfo().build())
                 .householdMemberSearch(HouseholdMemberSearch.builder()
