@@ -18,13 +18,13 @@ public class UpdateQueryBuilder implements GenericQueryBuilder{
         return paramsMap;
     }
     @Override
-    public String build(Object object) throws QueryBuilderException {
+    public String build(String schemaTemplate, Object object) throws QueryBuilderException {
         StringBuilder queryStringBuilder = null;
         try {
             String tableName = GenericQueryBuilder.getTableName(object.getClass());
             List<String> fieldsToUpdate = GenericQueryBuilder.getFieldsWithCondition(object,  QueryFieldChecker.isNotNull, paramsMap);
             List<String> fieldsToUpdateWith = GenericQueryBuilder.getFieldsWithCondition(object,  QueryFieldChecker.isAnnotatedWithUpdateBy, paramsMap);
-            queryStringBuilder = GenericQueryBuilder.generateQuery(GenericQueryBuilder.updateQueryTemplate(tableName), fieldsToUpdate, fieldsToUpdateWith);
+            queryStringBuilder = GenericQueryBuilder.generateQuery(GenericQueryBuilder.updateQueryTemplate(schemaTemplate, tableName), fieldsToUpdate, fieldsToUpdateWith);
         } catch (Exception exception) {
             throw new QueryBuilderException(exception.getMessage());
         }
