@@ -15,6 +15,7 @@ import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.Map;
 
+import static org.egov.common.utils.MultiStateInstanceUtil.SCHEMA_REPLACE_STRING;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -28,11 +29,11 @@ class GenericQueryBuilderTest {
                 .dummyString("some-string")
                 .dummyInt(1)
                 .build();
-        String expectedQuery = "SELECT * FROM dummyData WHERE " +
+        String expectedQuery = "SELECT * FROM {schema}.dummyData WHERE " +
                 "dummyString=:dummyString AND dummyInt=:dummyInt";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
 
         assertEquals(expectedQuery, actualQuery);
     }
@@ -46,11 +47,11 @@ class GenericQueryBuilderTest {
         DummyData data = DummyData.builder()
                 .dummyStringList(strings)
                 .build();
-        String expectedQuery = "SELECT * FROM dummyData WHERE " +
+        String expectedQuery = "SELECT * FROM {schema}.dummyData WHERE " +
                 "dummyStringList IN (:dummyStringList)";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
         Map<String, Object> paramMap = queryBuilder.getParamsMap();
 
         assertEquals(expectedQuery, actualQuery);
@@ -69,11 +70,11 @@ class GenericQueryBuilderTest {
                 .dummyStringList(strings)
                 .dummyIntegerList(ints)
                 .build();
-        String expectedQuery = "SELECT * FROM dummyData WHERE " +
+        String expectedQuery = "SELECT * FROM {schema}.dummyData WHERE " +
                 "dummyStringList IN (:dummyStringList)";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
         Map<String, Object> paramMap = queryBuilder.getParamsMap();
 
         assertEquals(expectedQuery, actualQuery);
@@ -88,10 +89,10 @@ class GenericQueryBuilderTest {
         DummyData data = DummyData.builder()
                 .dummyStringList(strings)
                 .build();
-        String expectedQuery = "SELECT * FROM dummyData";
+        String expectedQuery = "SELECT * FROM {schema}.dummyData";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
 
         assertEquals(expectedQuery, actualQuery);
     }
@@ -102,10 +103,10 @@ class GenericQueryBuilderTest {
         DummyData data = DummyData.builder()
                 .dummyStringList(null)
                 .build();
-        String expectedQuery = "SELECT * FROM dummyData";
+        String expectedQuery = "SELECT * FROM {schema}.dummyData";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
 
         assertEquals(expectedQuery, actualQuery);
     }
@@ -121,11 +122,11 @@ class GenericQueryBuilderTest {
                 .dummyPrimitiveInt(12)
                 .dummyPrimitiveFloat(232.2f)
                 .build();
-        String expectedQuery = "SELECT * FROM dummyData WHERE " +
+        String expectedQuery = "SELECT * FROM {schema}.dummyData WHERE " +
                 "dummyString=:dummyString AND dummyInt=:dummyInt";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
 
         assertEquals(expectedQuery, actualQuery);
     }
@@ -135,10 +136,10 @@ class GenericQueryBuilderTest {
     void shouldNotUseWhereClauseWhenPropertiesAreSetToNullSelectQuery() throws QueryBuilderException {
         DummyData data = DummyData.builder()
                 .build();
-        String expectedQuery = "SELECT * FROM dummyData";
+        String expectedQuery = "SELECT * FROM {schema}.dummyData";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
 
         assertEquals(expectedQuery, actualQuery);
     }
@@ -150,10 +151,10 @@ class GenericQueryBuilderTest {
                 .dummyString("TEST123")
                 .dummyAddress(DummyAddress.builder().addressString("123").build())
                 .build();
-        String expectedQuery = "SELECT * FROM dummyData WHERE dummyString=:dummyString AND addressString=:addressString";
+        String expectedQuery = "SELECT * FROM {schema}.dummyData WHERE dummyString=:dummyString AND addressString=:addressString";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
 
         assertEquals(expectedQuery, actualQuery);
     }
@@ -168,10 +169,10 @@ class GenericQueryBuilderTest {
                         .addressString("123")
                         .dummyAmount(DummyAmount.builder().amount(123.0).currency("INR").build()).build())
                 .build();
-        String expectedQuery = "SELECT * FROM dummyData WHERE dummyString=:dummyString AND addressString=:addressString AND currency=:currency AND amount=:amount";
+        String expectedQuery = "SELECT * FROM {schema}.dummyData WHERE dummyString=:dummyString AND addressString=:addressString AND currency=:currency AND amount=:amount";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
 
         assertEquals(expectedQuery, actualQuery);
     }
@@ -186,10 +187,10 @@ class GenericQueryBuilderTest {
                         .builder()
                         .addressString("123").build())
                 .build();
-        String expectedQuery = "UPDATE dummyData SET dummyString=:dummyString , dummyInt=:dummyInt , addressString=:addressString WHERE dummyID=:dummyID";
+        String expectedQuery = "UPDATE {schema}.dummyData SET dummyString=:dummyString , dummyInt=:dummyInt , addressString=:addressString WHERE dummyID=:dummyID";
         UpdateQueryBuilder queryBuilder = new UpdateQueryBuilder();
 
-        String actualQuery = queryBuilder.build(data);
+        String actualQuery = queryBuilder.build(SCHEMA_REPLACE_STRING, data);
 
         assertEquals(expectedQuery, actualQuery);
     }
@@ -201,10 +202,10 @@ class GenericQueryBuilderTest {
                 .dummyString("some-string")
                 .dummyID(1)
                 .build();
-        String expectedQuery = "UPDATE dummyData SET dummyString=:dummyString , dummyInt=:dummyInt , addressString=:addressString WHERE dummyID=:dummyID";
+        String expectedQuery = "UPDATE {schema}.dummyData SET dummyString=:dummyString , dummyInt=:dummyInt , addressString=:addressString WHERE dummyID=:dummyID";
         SelectQueryBuilder queryBuilder = new SelectQueryBuilder();
 
-        assertThrows(QueryBuilderException.class, ()-> queryBuilder.build(data));
+        assertThrows(QueryBuilderException.class, ()-> queryBuilder.build(SCHEMA_REPLACE_STRING, data));
     }
 
     @Data
