@@ -21,10 +21,7 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyBoolean;
-import static org.mockito.ArgumentMatchers.anyList;
-import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
@@ -86,7 +83,11 @@ class ProductVariantServiceSearchTest {
         ProductVariantSearchRequest productVariantSearchRequest = ProductVariantSearchRequest.builder()
                 .productVariant(productVariantSearch).requestInfo(RequestInfoTestBuilder.builder()
                         .withCompleteRequestInfo().build()).build();
-        when(productVariantRepository.findById(anyList(), anyBoolean())).thenReturn(productVariants);
+        when(productVariantRepository.findById(
+                isNull(),
+                eq(Collections.singletonList("ID101")),
+                eq(true)
+        )).thenReturn(productVariants);
 
         List<ProductVariant> productVariants = productVariantService.search(productVariantSearchRequest,
                 10, 0, null, null, true);
