@@ -1,5 +1,6 @@
 package org.egov.product.service;
 
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.models.product.ApiOperation;
 import org.egov.common.models.product.ProductVariant;
 import org.egov.common.models.product.ProductVariantRequest;
@@ -24,9 +25,7 @@ import java.util.stream.Collectors;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyList;
-import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -65,18 +64,18 @@ class ProductVariantServiceUpdateTest {
         lenient().when(productConfiguration.getUpdateProductVariantTopic()).thenReturn("update-topic");
     }
 
-    private void mockFindById() {
-        when(productVariantRepository.findById(productVariantIds)).thenReturn(request.getProductVariant());
+    private void mockFindById() throws InvalidTenantIdException {
+        when(productVariantRepository.findById(anyString(), eq(productVariantIds))).thenReturn(request.getProductVariant());
     }
 
-    private void mockValidateProuctId() {
-        when(productService.validateProductId(any(List.class)))
+    private void mockValidateProuctId() throws InvalidTenantIdException {
+        when(productService.validateProductId(anyString(), any(List.class)))
                 .thenReturn(Collections.singletonList("some-product-id"));
     }
 
     @Test
     @DisplayName("should update the lastModifiedTime in the result")
-    void shouldUpdateTheLastModifiedTimeInTheResult() {
+    void shouldUpdateTheLastModifiedTimeInTheResult() throws InvalidTenantIdException {
         Long time = request.getProductVariant().get(0).getAuditDetails().getLastModifiedTime();
         mockValidateProuctId();
         mockFindById();
@@ -89,7 +88,7 @@ class ProductVariantServiceUpdateTest {
 
     @Test
     @DisplayName("should update the row version in the result")
-    void shouldUpdateTheRowVersionInTheResult() {
+    void shouldUpdateTheRowVersionInTheResult() throws InvalidTenantIdException {
         Integer rowVersion = request.getProductVariant().get(0).getRowVersion();
         mockValidateProuctId();
         mockFindById();
@@ -102,47 +101,47 @@ class ProductVariantServiceUpdateTest {
 
     @Test
     @DisplayName("should check if the request has valid product ids")
-    void shouldCheckIfTheRequestHasValidProductIds() {
-        when(productService.validateProductId(any(List.class))).thenReturn(Collections
+    void shouldCheckIfTheRequestHasValidProductIds() throws InvalidTenantIdException {
+        when(productService.validateProductId(anyString(), any(List.class))).thenReturn(Collections
                 .singletonList(request.getProductVariant().get(0).getProductId()));
         mockFindById();
 
         productVariantService.update(request);
 
-        verify(productService, times(1)).validateProductId(any(List.class));
+        verify(productService, times(1)).validateProductId(anyString(), any(List.class));
     }
 
     @Test
     @DisplayName("should throw exception for any invalid product id")
     void shouldThrowExceptionForAnyInvalidProductId() throws Exception {
-        when(productService.validateProductId(any(List.class))).thenReturn(Collections.emptyList());
+        when(productService.validateProductId(anyString(), any(List.class))).thenReturn(Collections.emptyList());
 
         assertThrows(CustomException.class, () -> productVariantService.update(request));
     }
 
     @Test
     @DisplayName("should fetch existing records using id")
-    void shouldFetchExistingRecordsUsingId() {
+    void shouldFetchExistingRecordsUsingId() throws InvalidTenantIdException {
         mockValidateProuctId();
         mockFindById();
 
         productVariantService.update(request);
 
-        verify(productVariantRepository, times(1)).findById(anyList());
+        verify(productVariantRepository, times(1)).findById(anyString(), anyList());
     }
 
     @Test
     @DisplayName("should throw exception if fetched records count doesn't match the count in request")
-    void shouldThrowExceptionIfFetchedRecordsCountDoesntMatchTheCountInRequest() {
+    void shouldThrowExceptionIfFetchedRecordsCountDoesntMatchTheCountInRequest() throws InvalidTenantIdException {
         mockValidateProuctId();
-        when(productVariantRepository.findById(anyList())).thenReturn(Collections.emptyList());
+        when(productVariantRepository.findById(anyString(), anyList())).thenReturn(Collections.emptyList());
 
         assertThrows(CustomException.class, () -> productVariantService.update(request));
     }
 
     @Test
     @DisplayName("should send the updates to kafka")
-    void shouldSendTheUpdatesToKafka() {
+    void shouldSendTheUpdatesToKafka() throws InvalidTenantIdException {
         mockValidateProuctId();
         mockFindById();
         when(productVariantRepository.save(anyList(), anyString())).thenReturn(request.getProductVariant());
@@ -154,13 +153,13 @@ class ProductVariantServiceUpdateTest {
 
     @Test
     @DisplayName("Should throw exception for row versions mismatch")
-    void shouldThrowExceptionIfRowVersionIsNotSimilar() {
+    void shouldThrowExceptionIfRowVersionIsNotSimilar() throws InvalidTenantIdException {
         ProductVariant productVariant = ProductVariantTestBuilder.builder().withId().build();
         productVariant.setRowVersion(123);
         ProductVariantRequest productVariantRequest = ProductVariantRequestTestBuilder.builder()
                 .withOneProductVariantHavingId().build();
         mockValidateProuctId();
-        when(productVariantRepository.findById(productVariantIds))
+        when(productVariantRepository.findById(anyString(), eq(productVariantIds)))
                 .thenReturn(Collections.singletonList(productVariant));
 
         assertThrows(Exception.class, () -> productVariantService.update(productVariantRequest));
