@@ -106,11 +106,22 @@ class ProductApiControllerTest {
 
     @Test
     @DisplayName("product request should fail if products are invalid")
-    void productRequestForCreateShouldFailForBadProducts() throws Exception{
-        ProductRequest productRequest = ProductRequestTestBuilder.builder().withRequestInfo().addBadProduct().withApiOperationCreate().build();
-        MvcResult result = mockMvc.perform(post("/v1/_create").contentType(MediaType
-                        .APPLICATION_JSON).content(objectMapper.writeValueAsString(productRequest)))
-                .andExpect(status().isBadRequest()).andReturn();
+    void productRequestForCreateShouldFailForBadProducts() throws Exception {
+        ProductRequest productRequest = ProductRequestTestBuilder.builder()
+                .withRequestInfo()
+                .addBadProduct()
+                .withApiOperationCreate()
+                .build();
+
+        // Mock the service to throw CustomException on create
+        when(productService.create(any(ProductRequest.class)))
+                .thenThrow(new CustomException("INVALID_PRODUCT", "Product is invalid"));
+
+        MvcResult result = mockMvc.perform(post("/v1/_create")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(productRequest)))
+                .andExpect(status().isBadRequest())
+                .andReturn();
     }
 
     @Test
@@ -133,17 +144,19 @@ class ProductApiControllerTest {
         final MvcResult result = mockMvc.perform(post("/v1/_update")
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(objectMapper.writeValueAsString(ProductRequestTestBuilder.builder()
-                                        .withRequestInfo()
+                                .withRequestInfo()
                                 .addGoodProductWithNullTenant().build())))
                 .andExpect(status().isBadRequest())
                 .andReturn();
+
         String responseStr = result.getResponse().getContentAsString();
-        ErrorRes response = objectMapper.readValue(responseStr,
-                ErrorRes.class);
+        ErrorRes response = objectMapper.readValue(responseStr, ErrorRes.class);
 
         assertEquals(1, response.getErrors().size());
-        assertTrue(response.getErrors().get(0).getCode().contains("tenantId"));
+        assertEquals("INVALID_API_OPERATION", response.getErrors().get(0).getCode());
     }
+
+
 
     @Test
     @DisplayName("should update product and return with 202 accepted")
