@@ -190,7 +190,7 @@ public class StockService {
             List<String> ids = (List<String>) ReflectionUtils.invokeMethod(getIdMethod(Collections
                             .singletonList(stockSearchRequest.getStock())),
                     stockSearchRequest.getStock());
-            List<Stock> stocks = stockRepository.findById(ids, includeDeleted, idFieldName).stream()
+            List<Stock> stocks = stockRepository.findById(tenantId, ids, includeDeleted, idFieldName).stream()
                     .filter(lastChangedSince(lastChangedSince))
                     .filter(havingTenantId(tenantId))
                     .filter(includeDeleted(includeDeleted))
