@@ -6,10 +6,12 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 import lombok.extern.slf4j.Slf4j;
+import org.egov.common.exception.InvalidTenantIdException;
 import org.egov.common.models.Error;
 import org.egov.common.models.stock.StockReconciliation;
 import org.egov.common.models.stock.StockReconciliationBulkRequest;
 import org.egov.common.models.stock.StockReconciliationSearch;
+import org.egov.common.utils.CommonUtils;
 import org.egov.common.validator.Validator;
 import org.egov.stock.repository.StockReconciliationRepository;
 import org.springframework.core.annotation.Order;
@@ -19,6 +21,7 @@ import org.springframework.util.StringUtils;
 
 import static org.egov.common.utils.CommonUtils.notHavingErrors;
 import static org.egov.common.utils.CommonUtils.populateErrorDetails;
+import static org.egov.common.utils.ValidatorUtils.getErrorForInvalidTenantId;
 import static org.egov.common.utils.ValidatorUtils.getErrorForUniqueEntity;
 
 /**
@@ -53,7 +56,7 @@ public class SrExistentEntityValidator implements Validator<StockReconciliationB
     public Map<StockReconciliation, List<Error>> validate(StockReconciliationBulkRequest request) {
         // Map to hold StockReconciliation entities and their error details
         Map<StockReconciliation, List<Error>> errorDetailsMap = new HashMap<>();
-
+        String tenantId = CommonUtils.getTenantId(request.getStockReconciliation());
         // Get the list of StockReconciliation entities from the request
         List<StockReconciliation> entities = request.getStockReconciliation();
 
@@ -75,17 +78,28 @@ public class SrExistentEntityValidator implements Validator<StockReconciliationB
 
         // Check if the list of client reference IDs is not empty
         if (!CollectionUtils.isEmpty(clientReferenceIdList)) {
-            // Query the repository to find existing StockReconciliation entities with the given client reference IDs
-            List<String> existingClientReferenceIds =
-                    stockReconciliationRepository.validateClientReferenceIdsFromDB(clientReferenceIdList, Boolean.TRUE);
 
-            // For each existing client reference ID, add an error to the map for the corresponding StockReconciliation entity
-            existingClientReferenceIds.forEach(clientReferenceId -> {
-                // Get a predefined error object for unique entity validation
-                Error error = getErrorForUniqueEntity();
-                // Populate error details for the individual StockReconciliation entity associated with the client reference ID
-                populateErrorDetails(map.get(clientReferenceId), error, errorDetailsMap);
-            });
+            try {
+                // Query the repository to find existing StockReconciliation entities with the given client reference IDs
+                List<String> existingClientReferenceIds =
+                        stockReconciliationRepository.validateClientReferenceIdsFromDB(tenantId, clientReferenceIdList, Boolean.TRUE);
+
+                // For each existing client reference ID, add an error to the map for the corresponding StockReconciliation entity
+                existingClientReferenceIds.forEach(clientReferenceId -> {
+                    // Get a predefined error object for unique entity validation
+                    Error error = getErrorForUniqueEntity();
+                    // Populate error details for the individual StockReconciliation entity associated with the client reference ID
+                    populateErrorDetails(map.get(clientReferenceId), error, errorDetailsMap);
+                });
+            } catch (InvalidTenantIdException exception) {
+                entities.stream().forEach(stockReconciliation -> {
+                    Error error  = getErrorForInvalidTenantId(tenantId, exception);
+                    // Populate error details for all entities in case of tenant ID validation failure
+                    populateErrorDetails(stockReconciliation, error, errorDetailsMap);
+                });
+
+            }
+
         }
 
         // Return the map containing StockReconciliation entities and their associated error details
